% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckdata-package.R
\docType{package}
\name{ckdata-package}
\alias{ckdata}
\alias{ckdata-package}
\title{ckdata: Query the CourseKata Data Lake}
\description{
A thin R interface to the CourseKata analytical data lake: get lazy dbplyr table references into the lake and cache large query results to disk. The Amazon Athena connection is managed for you; queries are written with dbplyr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coursekata/ckdata-r}
}

}
\author{
\strong{Maintainer}: Adam Blake \email{adam@ablake.dev}

Authors:
\itemize{
  \item Adam Blake \email{adam@ablake.dev}
}

}
\keyword{internal}
